Renamed InLobbyView.cs to TopInfoPanelController.cs
- Added doc comment.
- Removed old unused code from TopInfoPanelController.cs and InLobbyController.cs
- Renamed the reference variable to TopInfoPanel in InLobbyController.cs and reconnected reference in Battle Popup prefab.
