+++ Assets/MenuUi/Scripts/Lobby/InLobby/InLobbyController.cs
@@ -27,7 +27,7 @@ namespace MenuUi.Scripts.Lobby.InLobby
     /// </summary>
     public class InLobbyController : AltMonoBehaviour
     {
-        [SerializeField] private InLobbyView _view;
+        [SerializeField] private TopInfoPanelController _topInfoPanel;
         [SerializeField] private SelectedCharactersPopup _selectedCharactersPopup;
         [SerializeField] private GameObject _popupContents;
         [SerializeField] private BattlePopupPanelManager _roomSwitcher;
@@ -36,10 +36,6 @@ namespace MenuUi.Scripts.Lobby.InLobby
 
         private void Awake()
         {
-            //_view.CharacterButtonOnClick = CharacterButtonOnClick;
-            //_view.RoomButtonOnClick = RoomButtonOnClick;
-            //_view.RaidButtonOnClick = RaidButtonOnClick;
-            //_view.QuickGameButtonOnClick = QuickGameButtonOnClick;
             SignalBus.OnBattlePopupRequested += TryOpenWindow;
         }
 
@@ -64,9 +60,9 @@ namespace MenuUi.Scripts.Lobby.InLobby
                 // We need to disconnect from current region because it is not the same as in player settings.
                 PhotonLobby.Disconnect();
             }*/
-            _view.Reset();
+            _topInfoPanel.Reset();
             UpdateTitle();
-            _view.LobbyText = string.Empty;
+            _topInfoPanel.LobbyText = string.Empty;
             StartCoroutine(StartLobby(playerSettings.PlayerGuid, playerSettings.PhotonRegion));
         }
 
@@ -79,7 +75,7 @@ namespace MenuUi.Scripts.Lobby.InLobby
         {
             // Save region for later use because getting it is not cheap (b ut not very expensive either). 
             _currentRegion = PhotonRealtimeClient.CloudRegion != null ? PhotonRealtimeClient.CloudRegion : "";
-            _view.TitleText = $"{Application.productName} {PhotonRealtimeClient.GameVersion}";
+            _topInfoPanel.TitleText = $"{Application.productName} {PhotonRealtimeClient.GameVersion}";
         }
 
         private IEnumerator StartLobby(string playerGuid, string photonRegion)
@@ -114,21 +110,19 @@ namespace MenuUi.Scripts.Lobby.InLobby
                 yield return delay;
             }
             UpdateTitle();
-            _view.EnableButtons();
         }
 
         private void Update()
         {
             if (!PhotonRealtimeClient.InLobby && !PhotonRealtimeClient.InRoom)
             {
-                _view.LobbyText = "Wait";
+                _topInfoPanel.LobbyText = "Wait";
                 return;
             }
             UpdateTitle();
-            _view.EnableButtons();
             var playerCount = PhotonRealtimeClient.CountOfPlayers;
-            _view.LobbyText = $"Alue: {_currentRegion} : {PhotonRealtimeClient.GetPing()} ms";
-            _view.PlayerCountText = $"Pelaajia online: {playerCount}";
+            _topInfoPanel.LobbyText = $"Alue: {_currentRegion} : {PhotonRealtimeClient.GetPing()} ms";
+            _topInfoPanel.PlayerCountText = $"Pelaajia online: {playerCount}";
         }
 
         /*public void OnDisconnected(DisconnectCause cause)
